add testing utils
