+++ test/ast.test.js
@@ -1,37 +1,6 @@
 import { describe, expect, it } from 'vitest'
-import { tokenize, SugarHigh } from '../lib'
-
-function getTypeName(token) {
-  return SugarHigh.TokenTypes[token[0]]
-}
-
-function getTokenValues(tokens) {
-  return tokens.map((tk) => tk[1])
-}
-
-function mergeSpaces(str) {
-  return str.trim().replace(/^[\s]{2,}$/g, ' ')
-}
-
-function filterSpaces(arr) {
-  return arr
-    .map(t => mergeSpaces(t))
-    .filter(Boolean)
-}
-
-function extractTokenValues(tokens) {
-  return filterSpaces(getTokenValues(tokens))
-}
-
-function getTokenArray(tokens) {
-  return tokens.map((tk) => [tk[1], getTypeName(tk)]);
-}
-
-function extractTokenArray(tokens) {
-  return tokens
-    .map((tk) => [mergeSpaces(tk[1]), getTypeName(tk)])
-    .filter(([_, type]) => type !== 'space' && type !== 'break')
-}
+import { tokenize } from '../lib'
+import { extractTokenArray, extractTokenValues } from './testing-utils'
 
 describe('function calls', () => {
   it('dot catch should not be determined as keyword', () => {
